refactor: remove comment work from bigpic
